+++ bot/core/core.js
@@ -188,19 +188,20 @@ module.exports.acceptBotInteraction = (bot) => {
         roleUpdate(bot, roleBefore, roleAfter);
     });
 
-    bot.on('guildBanAdd', (guild, user) => {
-        guildBanAdd(bot, guild, user);
+    bot.on('guildBanAdd', (guildBan) => {
+        guildBanAdd(bot, guildBan);
     });
 
-    bot.on('guildBanRemove', (guild, user) => {
-        guildBanRemove(bot, guild, user);
+    bot.on('guildBanRemove', (guildBan) => {
+        guildBanRemove(bot, guildBan);
     });
 };
 
 module.exports.fetchCache = async (bot) => {
     return new Promise(async (resolve, reject) => {
         try {
-            console.time('Fetching guilds in:');
+            console.log(`Starting to fetch ${bot.guilds.cache.size} guilds...`)
+            console.time(`Fetching guilds in:`);
             const guilds = await bot.guilds.fetch();
             console.timeEnd('Fetching guilds in:');