+++ src/events/notfifier/yt_notifier.js
@@ -1,6 +1,7 @@
 const request = new (require('rss-parser'))();
 const { errorhandler } = require('../../../utils/functions/errorhandler/errorhandler');
 const guildUploads = require('../../db/Models/tables/guildUploads.model');
+const yt = require('ytdl-core');
 
 module.exports.handleUploads = async ({ bot }) => {
     console.info('🔎 Youtube upload handler started');
@@ -16,9 +17,7 @@ module.exports.handleUploads = async ({ bot }) => {
                 });
                 return false;
             });
-
         if (uploads.length === 0) return false;
-
         for (let i in uploads) {
             if (uploads[i].channel_id) {
                 request
@@ -30,6 +29,33 @@ module.exports.handleUploads = async ({ bot }) => {
 
                         const videoAlreadyExists = uploadedVideos.includes(feed.items[0].link);
                         if (videoAlreadyExists) return;
+
+                        const isALiveVideoOrPremiere = await yt
+                            .getInfo(feed.items[0].link)
+                            .then(async (info) => {
+                                return info.videoDetails.liveBroadcastDetails;
+                            })
+                            .catch((err) => {
+                                errorhandler({
+                                    err,
+                                    fatal: true,
+                                });
+                                return false;
+                            });
+
+                        let premiereStartsIn;
+                        if (isALiveVideoOrPremiere) {
+                            const isLiveNow = isALiveVideoOrPremiere.isLiveNow;
+                            if (!isLiveNow) return;
+
+                            const year = isALiveVideoOrPremiere.startTimestamp.substring(0, 4);
+                            const month = isALiveVideoOrPremiere.startTimestamp.substring(5, 7) - 1;
+                            const day = isALiveVideoOrPremiere.startTimestamp.substring(8, 10);
+                            const hour =
+                                isALiveVideoOrPremiere.startTimestamp.substring(11, 13) - 1;
+                            const date = new Date(year, month, day, hour);
+                            premiereStartsIn = date.getTime() / 1000;
+                        }
 
                         uploadedVideos.push(feed.items[0].link);
 
@@ -76,7 +102,11 @@ module.exports.handleUploads = async ({ bot }) => {
                                             : `<@&${uploads[i].pingrole}> `
                                         : '') +
                                     feed.items[0].title +
-                                    ` ${feed.items[0].link}`,
+                                    ` ${feed.items[0].link} ${
+                                        isALiveVideoOrPremiere
+                                            ? `\n**Premiere starts in <t:${premiereStartsIn}:R>**`
+                                            : ''
+                                    }`,
                             })
                             .catch((err) => {});
 
@@ -89,6 +119,7 @@ module.exports.handleUploads = async ({ bot }) => {
                         errorhandler({
                             message: 'Youtube request run into Timeout.',
                             fatal: err.errno === 'ECONNREFUSED' ? false : true,
+                            err,
                         });
                         return false;
                     });