+++ src/slash_commands/index.js
@@ -25,6 +25,8 @@ module.exports.handleSlashCommands = async ({ main_interaction, bot }) => {
         'automod',
         'modroles',
         'log',
+        'autoblacklist',
+        'joinroles',
     ];
     const level = ['rank', 'leaderboard', 'givexp', 'removexp'];
     const utils = [